+++ test/src/androidTest/kotlin/io/realm/shared/SchemaTests.kt
@@ -24,10 +24,23 @@ import test.link.Child
 import test.link.Parent
 import kotlin.reflect.KClass
 import kotlin.test.Test
+import kotlin.test.assertContentEquals
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
 class SchemaTests {
+
+    @Test
+    fun defaultConfig() {
+        val config = RealmConfiguration.defaultConfig(schema=setOf(Sample::class))
+        assertEquals(setOf(Sample::class), config.schema)
+        assertEquals<Map<KClass<out RealmObject>, io.realm.internal.RealmObjectCompanion>>(
+            mapOf(
+                Sample::class to (Sample as io.realm.internal.RealmObjectCompanion)
+            ),
+            config.companionMap
+        )
+    }
 
     @Test
     fun usingNamedArgument() {